requirements.txt
blinker==1.9.0
click==8.1.7
requests==2.32.3
Flask==3.1.0
itsdangerous==2.2.0
Werkzeug==3.1.3
python-dotenv==1.0.1
yt-dlp==2024.12.13
